requirements.txt
absl-py==0.9.0
alabaster==0.7.12
anaconda-client==1.7.2
anaconda-navigator==1.9.7
asn1crypto==1.0.1
astor==0.8.1
astroid==1.6.5
atomicwrites==1.4.0
attrs==19.2.0
Babel==2.7.0
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.8.0
bleach==3.1.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
clyent==1.2.2
conda==4.7.12
conda-build==3.18.10
conda-package-handling==1.6.0
conda-verify==3.4.2
configparser==4.0.2
contextlib2==0.6.0
cryptography==2.7
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
docutils==0.15.2
entrypoints==0.3
enum34==1.1.6
filelock==3.0.12
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.17.1
futures==3.3.0
gast==0.3.3
glob2==0.7
google-pasta==0.1.8
grpcio==1.27.2
h5py==2.10.0
icc-rt==2019.0
idna==2.8
imagesize==1.1.0
importlib-metadata==2.0.0
innvestigate==1.0.8
intel-openmp==2019.0
ipaddress==1.0.22
ipykernel==4.10.0
ipython==5.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.15.1
Jinja2==2.10.3
jsonschema==3.0.2
jupyter-client==5.3.4
jupyter-core==4.6.0
jupyterlab==0.33.11
jupyterlab-launcher==0.11.2
Keras==2.2.4
Keras-Applications==1.0.8
keras-contrib==2.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
libarchive-c==2.8
lief==0.9.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==2.2.5
mccabe==0.6.1
mistune==0.8.4
mkl==2019.0
mock==3.0.5
more-itertools==5.0.0
navigator-updater==0.2.1
nbconvert==5.6.0
nbformat==4.4.0
notebook==5.7.8
numpy==1.16.0
numpydoc==0.9.1
olefile==0.46
opencv-python==4.2.0.32
packaging==19.2
pandas==0.24.2
pandocfilters==1.4.2
parso==0.5.1
pathlib2==2.3.5
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.0
pkginfo==1.5.0.1
pluggy==0.13.1
progressbar==2.5
prometheus-client==0.7.1
prompt-toolkit==1.0.15
protobuf==3.11.3
psutil==5.6.3
ptyprocess==0.6.0
py==1.9.0
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pyflakes==2.1.1
Pygments==2.4.2
pylint==1.9.2
pyOpenSSL==19.0.0
pyparsing==2.4.2
pyrsistent==0.15.4
PySocks==1.7.1
pytest==4.6.11
python-dateutil==2.8.0
python-util==1.2
pytz==2019.3
PyYAML==5.1.2
pyzmq==18.1.0
QtAwesome==0.6.0
qtconsole==4.5.5
QtPy==1.9.0
requests==2.22.0
rope==0.14.0
ruamel-yaml==0.15.46
scandir==1.10.0
scikit-learn==0.20.4
scipy==1.2.2
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
snowballstemmer==2.0.0
soupsieve==1.9.3
Sphinx==1.8.5
sphinxcontrib-websupport==1.1.2
spyder==3.3.6
spyder-kernels==0.5.2
subprocess32==3.5.4
tbb==2019.0
tbb4py==2019.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tornado==5.1.1
tqdm==4.36.1
traitlets==4.3.3
typing==3.7.4.1
urllib3==1.24.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
wrapt==1.11.2
wurlitzer==1.0.3
zipp==1.2.0
